Fix dir.walk to function as specified in docs
The documentation says that the list of subdirectories can be modified
to avoid recursing down into a folder, but the function currently
doesn't support that. This change makes the actual functionality match
the documentation.

When functioning as specified in the docs, the bottom_up implementation
is different enough from the default top-down implementation that I just
separated them.